# src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8090
  description: Generated server url
...
paths:
  /api/actors:
    get:
      tags:
      - actor-entity-controller
      description: get-actor
      operationId: getCollectionResource-actor-get_1
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          minimum: 1
          type: integer
          default: 20
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/PagedModelEntityModelActor'
            application/x-spring-data-compact+json:
              schema:
                $ref: '#/components/schemas/PagedModelEntityModelActor'
            text/uri-list:
              schema:
                type: string
    post:
      tags:
      - actor-entity-controller
      description: create-actor
      operationId: postCollectionResource-actor-post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorRequestBody'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelActor'
  /api/actors/search/name:
    get:
      tags:
      - actor-search-controller
      operationId: executeSearch-actor-get
      parameters:
      - name: q
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelActor'
        "404":
          description: Not Found
  /api/actors/{id}:
    get:
      tags:
      - actor-entity-controller
      description: get-actor
      operationId: getItemResource-actor-get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelActor'
        "404":
          description: Not Found
    put:
      tags:
      - actor-entity-controller
      description: update-actor
      operationId: putItemResource-actor-put
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorRequestBody'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelActor'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelActor'
        "204":
          description: No Content
    delete:
      tags:
      - actor-entity-controller
      description: delete-actor
      operationId: deleteItemResource-actor-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - actor-entity-controller
      description: patch-actor
      operationId: patchItemResource-actor-patch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorRequestBody'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelActor'
        "204":
          description: No Content
...
  /api/titles/{id}/actor:
    get:
      tags:
      - title-property-reference-controller
      description: get-actor-by-title-Id
      operationId: followPropertyReference-title-get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelActor'
            text/uri-list:
              schema:
                type: string
        "404":
          description: Not Found
    put:
      tags:
      - title-property-reference-controller
      description: update-actor-by-title-Id
      operationId: createPropertyReference-title-put
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelActor'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelActor'
        "204":
          description: No Content
    delete:
      tags:
      - title-property-reference-controller
      description: delete-actor-by-title-Id
      operationId: deletePropertyReference-title-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - title-property-reference-controller
      description: patch-actor-by-title-Id
      operationId: createPropertyReference-title-patch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelActor'
        "204":
          description: No Content
  /api/titles/{id}/actor/{propertyId}:
    get:
      tags:
      - title-property-reference-controller
      description: get-actor-by-title-Id
      operationId: followPropertyReference-title-get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelActor'
        "404":
          description: Not Found
    delete:
      tags:
      - title-property-reference-controller
      description: delete-actor-by-title-Id
      operationId: deletePropertyReferenceId-title-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
  /api/titles/{id}/category:
    get:
      tags:
      - title-property-reference-controller
      description: get-category-by-title-Id
      operationId: followPropertyReference-title-get_2_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelCategory'
            text/uri-list:
              schema:
                type: string
        "404":
          description: Not Found
    put:
      tags:
      - title-property-reference-controller
      description: update-category-by-title-Id
      operationId: createPropertyReference-title-put_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelCategory'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelCategory'
        "204":
          description: No Content
    delete:
      tags:
      - title-property-reference-controller
      description: delete-category-by-title-Id
      operationId: deletePropertyReference-title-delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - title-property-reference-controller
      description: patch-category-by-title-Id
      operationId: createPropertyReference-title-patch_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelCategory'
        "204":
          description: No Content
  /api/titles/{id}/category/{propertyId}:
    get:
      tags:
      - title-property-reference-controller
      description: get-category-by-title-Id
      operationId: followPropertyReference-title-get_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelCategory'
        "404":
          description: Not Found
    delete:
      tags:
      - title-property-reference-controller
      description: delete-category-by-title-Id
      operationId: deletePropertyReferenceId-title-delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
...
  /api/recommend/category/{id}:
    get:
      tags:
      - recommend-controller
      operationId: bestInCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: top
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
...
components:
  schemas:
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
...
    CollectionModelObject:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            objects:
              type: array
              items:
                type: object
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    CollectionModelCategory:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            categories:
              type: array
              items:
                $ref: '#/components/schemas/CategoryResponse'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    CollectionModelActor:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            actors:
              type: array
              items:
                $ref: '#/components/schemas/ActorResponse'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
...
    EntityModelActor:
      required:
      - name
      type: object
      properties:
        name:
          type: string
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    PagedModelEntityModelActor:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            actors:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelActor'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
        page:
          $ref: '#/components/schemas/PageMetadata'
    CollectionModelEntityModelActor:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            actors:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelActor'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
...
    ActorRequestBody:
      required:
      - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        title:
          type: array
          items:
            type: string
...
    ActorResponse:
      required:
      - name
      type: object
      properties:
        name:
          type: string
    CategoryResponse:
      required:
      - name
      type: object
      properties:
        name:
          type: string
...
  securitySchemes:
    netflixAuth:
      type: http
      in: header
      scheme: basic